+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 14.0.2
 * Add a new setting (default on) that gives unique letters across all types of actors. Example: Goblin A, Goblin B, Hobgoblin C, Hobgoblin D, Hobgoblin E.
+* Add a new setting (default on) to optionally include unique creatures (singletons) in the lettering.
 
 # 14.0.1